requirements.txt
# Core OSC communication
python-osc>=1.8.0

# Voice input - Speech-to-Text
openai-whisper>=20231117

# Audio capture and processing
pyaudio>=0.2.13
sounddevice>=0.4.6  # Audio feedback playback
soundfile>=0.12.1  # Audio file I/O
webrtcvad>=2.0.10  # Voice Activity Detection
numpy>=1.24.0
scipy>=1.11.0  # Audio resampling in Whisper STT

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
word2number>=1.1  # Convert spoken numbers to digits (e.g., "one twenty" -> 120)

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
